Change njit to jit with explicit type for get_stress method

+++ fplib3_api4ase.py
@@ -323,11 +323,9 @@ class fp_GD_Calculator(Calculator):
         znucl = self.znucl
 
         if self.check_restart(atoms) or self._energy is None:
-            # write_vasp('input.vasp', atoms, direct=True)
             lat = atoms.cell[:]
             rxyz = atoms.get_positions()
-            # print("fp_energy lat=\n", lat)
-            # print("fp_energy rxyz=\n", rxyz)
+            
             lat = np.array(lat, dtype = np.float64)
             rxyz = np.array(rxyz, dtype = np.float64)
             types = np.int32(types)
@@ -361,11 +359,9 @@ class fp_GD_Calculator(Calculator):
         znucl = self.znucl
 
         if self.check_restart(atoms) or self._forces is None:
-            # write_vasp('input.vasp', atoms, direct=True)
             lat = atoms.cell[:]
             rxyz = atoms.get_positions()
-            # print("fp_forces lat=\n", lat)
-            # print("fp_forces rxyz=\n", rxyz)
+            
             lat = np.array(lat, dtype = np.float64)
             rxyz = np.array(rxyz, dtype = np.float64)
             types = np.int32(types)
@@ -400,13 +396,10 @@ class fp_GD_Calculator(Calculator):
         znucl = self.znucl
 
         if self.check_restart(atoms) or self._stress is None:
-            # write_vasp('input.vasp', atoms, direct=True)
             lat = atoms.cell[:]
             rxyz = atoms.get_positions()
             pos = atoms.get_scaled_positions()
-            # print("fp_stress lat=\n", lat)
-            # print("fp_stress rxyz=\n", rxyz)
-            # print("fp_stress pos=\n", pos)
+            
             lat = np.array(lat, dtype = np.float64)
             rxyz = np.array(rxyz, dtype = np.float64)
             types = np.int32(types)
@@ -415,8 +408,7 @@ class fp_GD_Calculator(Calculator):
             nx = np.int32(nx)
             lmax = np.int32(lmax)
             cutoff = np.float64(cutoff)
-            # forces = self._forces
-            forces=atoms.get_forces()
+            forces = np.float64(atoms.get_forces())
             stress = fplib3.get_stress(lat, rxyz, forces)
             # stress = fplib3.get_stress(lat, rxyz, types, znucl,
             #                            contract = contract,
@@ -437,7 +429,6 @@ class fp_GD_Calculator(Calculator):
         types = self.types
         znucl = self.znucl
 
-        # write_vasp('input.vasp', atoms, direct=True)
         lat = atoms.cell[:]
         rxyz = atoms.get_positions()
         lat = np.array(lat, dtype = np.float64)
